backend/requirements.txt
djangorestframework~=3.14.0
django~=3.0
Pillow~=8.4.0
numpy~=1.23.5
tifffile~=2023.3.15
django-redis~=5.2.0
pybase64~=1.2.3
requests==2.26.0
gunicorn
tensorflow~=2.12.0
matplotlib~=3.7.1
keras~=2.12.0
PyYAML~=6.0
PySimpleGUI~=4.60.5